UPSTREAM: media: cec-notifier: add cec_notifier_parse_hdmi_phandle helper
Add helper function to parse the DT for the hdmi-phandle property
and return the corresponding struct device pointer.

It takes care to avoid increasing the device refcount since all
we need is the device pointer. This pointer is used in the
notifier list as a key, but it is never accessed by the CEC driver.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Reported-by: Wen Yang <wen.yang99@zte.com.cn>
Acked-by: Wen Yang <wen.yang99@zte.com.cn>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit fbbd403b3286b4467f8b755efa0f10819cef9bba)

BUG=b:161871573
TEST=None

Signed-off-by: Sam McNally <sammc@chromium.org>
Change-Id: I3b7b2ebdad9ff1abdbdf02509d90e7d896739404
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2387384
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>