UPSTREAM: HID: input: fix assignment of .value
The value field is actually an array of .maxfield. We should assign the
correct number to the correct usage.

Not that we never encounter a device that requires this ATM, but better
have the proper code path.

Fixes: 2dc702c991e377 ("HID: input: use the Resolution Multiplier for
       high-resolution scrolling")
Cc: stable@vger.kernel.org  # v5.0+
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
(cherry picked from commit 39b3c3a5fbc5d744114e497d35bf0c12f798c134)

BUG=chromium:998168,chromium:888172
TEST=check MS Comfort Optical Mouse 3000

Change-Id: I866001d0878c2648630860dc1790cc03634295ef
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1778922
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>