Revert "CHROMIUM: iwl7000: use sync nmi in case of init flow failure"
This reverts commit c3c31a59611761c29365b8b041e34c3a9ba96f56.
Reason for revert: b/130651313
iwl_trans_sync_nmi assumes that MSI-X is enabled and tries to read a
register that is not part of iomem (CSR_MSIX_HW_INT_CAUSES_AD).
MSI-X is not enabled on StP2 devices.

BUG=b:130651313, chromium:951645
TEST=None. 4.19 platform (hatch) has MSI-X enabled, so it is not
possible to test this change (without taking the trouble to replace CcP2
with StP2 on hatch).

Change-Id: I3dfa7e3e3a54e2fd251efe2d4b852e16df9a48be
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1575099
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>