Revert "CHROMIUM: iwl7000: add sync_nmi to trans ops"
This reverts commit 67971af3c1015563cce66f54a9351fc4b7e10189.
Reason for revert: b/130651313
iwl_trans_sync_nmi assumes that MSI-X is enabled and tries to read a
register that is not part of iomem (CSR_MSIX_HW_INT_CAUSES_AD).
MSI-X is not enabled on StP2 devices.

BUG=b:130651313, chromium:951645
TEST=None. 4.19 platform is running HrP2/CcP2 both of which have MSI-X
enabled. One way to test this would be to replace CcP2 with StP2 on the
WHL Hatch proto, not going that route since we have enough time to fix
regressions on Hatch.

Change-Id: I2c82d8c2c058b54e38a10b3c26e74abce7ebc567
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1575100
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>