Revert "CHROMIUM: iwl7000: fix send hcmd timeout recovery flow"
This reverts commit 28e331e858a6c8c43f3d7cb7e1f1651638e01026.
Reason for revert: b/130651313
iwl_trans_sync_nmi assumes that MSI-X is enabled and tries to read a
register that is not part of iomem (CSR_MSIX_HW_INT_CAUSES_AD).
MSI-X is not enabled on StP2 devices.

BUG=b:130651313, chromium:951645
TEST=None. $BUG doesn't show up on 4.19 because 4.19 platforms are
running HrP2 or CcP2 which have MSI-X enabled. Taking this in with build
testing. One way to test would be to replace CcP2 with StP2 on WHL Hatch
proto.

Change-Id: I79c868288f53c0fbb9e3b9dd19ad7f51f5a95793
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1575101
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>