CHROMIUM: devfreq: rockchip: Add clk-rk3288-dmc driver
The driver provides an API to change the frequency of the DDR PLL.
DDR memory must not be accessed during a DDR frequency change. To
achieve this the following steps are involved:

- pause all CPUs except the primary one
- disable interrupts
- disable MMU
- execute code to change the DPLL rate and update controller parameters
  from SRAM
- resume execution from DDR, re-enable MMU, interrupts and CPUs

To avoid a janky display DDR frequency changes need to be synchronized
with display VBLANK events. The dmc_sync driver is used for this purpose.
Before a frequency change we obtain a deadline from dmc_sync by which the
rate change needs to be completed. During the steps before the actual
rate change the driver repeatedly performs checks and aborts the
operation if the deadline was exceeded.

This code is mostly copied from the downstream Chrome OS 3.14 tree
with the following changes:

- don't register the DMC clock with the Common Clock Framework (CCF).
  store a pointer to the instance struct in a static global, since it
  isn't associated with a clk object
- delete now unused CCF callbacks
- rk3288_dmcclk_round_rate()
  - adapted API
  - use fixed 24 MHz rate for IO_XIN24M
- rk3288_dmcclk_set_rate()
  - adapted API
  - round requested rate to available rate (formerly done by CCF)
  - do nothing if rate hasn't changed
  - store current frequency in existing dmc->cur_freq
- use 'service_bus' instead of 'noc' to be in line with
  the DT and the TRM
- use 'ret' macro in call_with_mmu_disabled (see 6ebbf2ce437b "ARM:
  convert all "mov.* pc, reg" to "bx reg" for ARMv6+" for reference)
- replaced BSYM() with badr macro (see 14327c662822 "ARM: replace
  BSYM() with badr assembly macro")
- moved from drivers/clk/rockchip/ to drivers/devfreq/rockchip/

BUG=chromium:941638
TEST=with CL:1574236 and CONFIG_ARM_RK3288_DMC_DEVFREQ=y on veyron:
  setting of min/max_freq in /sys/class/devfreq/devfreqN is
  reflected in clock 'dpll_ddr'
  (see /sys/kernel/debug/clk/clk_summary)

Change-Id: I5368dae651e50dbf9bffb8409f46ea41932f59d3
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1574234
Reviewed-by: Douglas Anderson <dianders@chromium.org>