CHROMIUM: platform/chrome: wilco_ec: Add charge scheduling backend
This patch adds the backend to support to control two
charge-scheduling policies managed by the EC, Advanced Charging
and Peak Shift.

Advanced Charging Mode maximizes the battery health by adjusting
the charging algorithm throughout the day. When enabled, the
system will use standard charging algorithms and other
techniques during non-work hours to maximize battery health.
During work hours, an express charge is used, allowing the
battery to be charged faster. For each day the time in which the
system will be most heavily used is specified by the start time
and the duration.

Peak Shift is power saving policy that minimizes AC usage during
the peak usage times of the day. For each weekday a start, end,
and charge_start time to run in Peak Shift mode can be set.
After the start time the system will run from the battery even
if the AC is attached, as long as the battery stays above a
specified threshold. After the end time the system will run from
AC (if attached) but will not charge the battery. The system
will again function normally (use AC and recharge the battery)
after the charge_start time.

BUG=b:128524060, b:125011171
TEST=With following patches this library works.

Change-Id: I39cde67294a315add09576fe870bc456b70bd43c
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1590689
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>