CHROMIUM: devfreq: clk-rk3288-dmc: Be less spammy
DDR frequency changes sometimes are aborted if they can't be completed
during a display VBLANK. In case of a timeout the DMC driver performs
a few retries. Currently an error message is logged on timeout failures,
which is a bit spammy since these timeouts are known to occur and a retry
mechanism is in place. Change the log level of these messages from ERROR
to DEBUG. If all retries fail the DMC driver still logs an error message.

BUG=chromium:941638
TEST=build for minnie

Change-Id: I6eb894fedc97f73a92e42eb2037d4a868c248b46
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1592006
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>