CHROMIUM: iwl7000: mvm: make wgds_tbl_idx signed
We use wgds_tbl_idx to store the return value of the call to
iwl_mvm_get_sar_geo_profile() and check if it's less than zero (an
error can be returned), but we used u32 instead of int, so the
condtion will never be true.  Fix this by making the variable an
integer.

Change-Id: Iac98fcea067cd6e736dad46561d35619476eb5de
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: f97e2620c58fc0902a9c9f6638b086224c33ad17