FROMGIT: drm/bridge/synopsys: dw-hdmi: Add "unwedge" for ddc bus
See the PhD thesis in the comments in this patch for details, but to
summarize this adds a hacky "unwedge" feature to the dw_hdmi i2c bus to
workaround what appears to be a hardware errata.  This relies on a
pinctrl entry to help change around muxing to perform the unwedge.

NOTE that the specific TV this was tested on was the "Samsung
UN40HU6950FXZA" and the specific port was the "STB" port.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20190502225336.206885-2-dianders@chromium.org
(cherry picked from commit 50f9495efe308eb25fd921ea7c8c8143ddeeae30
 git://anongit.freedesktop.org/drm-misc drm-misc-next)

BUG=chromium:971474
TEST=Test with monitor from partner lab

Change-Id: Ifaffa5b7cc6f143fa6f3b04dc6f001c528aa8e95
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1646941
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>