FROMLIST: mfd: Add support for the MediaTek MT6358 PMIC
This adds support for the MediaTek MT6358 PMIC. This is a
multifunction device with the following sub modules:

- Regulator
- RTC
- Codec
- Interrupt

It is interfaced to the host controller using SPI interface
by a proprietary hardware called PMIC wrapper or pwrap.
MT6358 MFD is a child device of the pwrap.

Signed-off-by: Hsin-Hsiung Wang <hsin-hsiung.wang@mediatek.com>

(am from https://patchwork.kernel.org/patch/10928149/)

BUG=b:126006378
TEST=build and boot to shell

Change-Id: Ia5e463bc048b58c140f87772d4dacbd24eb34c76
Signed-off-by: Claire Chang <tientzu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1598684
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>