CHROMIUM: devfreq: Fix rk3288_dmc / cpufreq interactions
When I switched my device over to the "schedutil" governor and it was
sitting on the sign in page (instead of the spinning login animation)
I started seeing _tons_ of devfreq timeouts spamming my console.

I did a tad bit of debugging and found that something about the way
rk3288_dmcfreq was trying to lock the CPU Frequency just wasn't
working.  Specifically I printed the frequency of "pll_apll" when I
was seeing a timeout and found that it was really low.

Instead of debugging why the old way of doing things didn't work
anymore, I decided to just switch rk3288_dmc to request a higher CPU
Frequency by using the cpufreq notifiers.  That's what everyone else
does and is the supported way.  It seems to work great.

As a side effect of this it turns out that we can easily override the
CPU Frequency that thermal throttling (and, actually, sysfs) imposed
on the system.  Thinking about this it actually seems like a
beneficial thing to do.  Sure we'll temporarily bump up to a really
high CPU when we're hot, but we'll very quickly bump back down.
...and overall we may save power because we won't get so many failed
transitions and we'll also spend less time with all CPUs spinning
waiting for the transition to be done.

BUG=chromium:941638
TEST=Stress the system and

Change-Id: I750ad080f3ac70c3aee8f85f174b77ad7a06a8e6
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1600369
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>