CHROMIUM: iwl7000: mvm: correctly fill the ac array in the iwl_mac_ctx_cmd
The indexes into the ac array in the iwl_mac_ctx_cmd are from the iwl_ac
enum and not the txfs.  The current code therefore puts the edca params
in the wrong indexes of the array, causing wrong priority for
data-streams of different ACs.
Fix this.

Note that this bug only occurs in NICs that use the new tx api, since in
the old tx api the txf number is equal to the corresponding ac in the
iwl_ac enum.

Change-Id: I7047bb5360fe24bcc275ef2f94c5bbb1e06e319d
Signed-off-by: Naftali Goldstein <naftali.goldstein@intel.com>
iwl7000-tree: 4192a86f956815e4984dcf6f12fc744d33e7b9ed