BACKPORT: FROMGIT: nvme-pci: Unblock reset_work on IO failure
The reset_work waits for queued IO to complete before setting the
controller to live. If any of these times out and requeues, we won't be
able to restart the controller because the reset_work is already running.

Flush all entered requests to a failed completion if a timeout occurs
in the connecting state, and ensure the controller can't transition to
the live state after we've unblocked it from waiting for completions.

Signed-off-by: Keith Busch <keith.busch@intel.com>
(cherry picked from commit 34baefb380c64da01f5e310a91d4890d1faab009
 git://git.kernel.org/pub/scm/linux/kernel/git/kbusch/linux.git
 nvme-power)

BUG=b:127593309,b:134195632,b:134781711
TEST=Tested with LiteOn CL1-3D256-Q11 NVMe.  DP/N 0TN2CC.  F/W 22301100

Change-Id: I54313849f3f337eb05831b16c7e719c52f81f17c
Signed-off-by: Rajat Jain <rajatja@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1630724
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>