CHROMIUM: iwl7000: pcie: increase the size of PCI dumps
Currently we dump only the first 64 bytes of the PCI config space,
which leaves out some important things, such as the base address
registers.

Increase it to 352 for the PCI device and to 524 for the rootport to
make sure we include everything we need.

Change-Id: Id359d6141d91f29d972f1fd5ebfc1d05e184b85c
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: b883e518d78dc7e60dfef039dcf93bb981ed10ca