BACKPORT: FROMGIT: spi: Allow SPI devices to request the pumping thread be realtime
Right now the only way to get the SPI pumping thread bumped up to
realtime priority is for the controller to request it.  However it may
be that the controller works fine with the normal priority but
communication to a particular SPI device on the bus needs realtime
priority.

Let's add a way for devices to request realtime priority when they set
themselves up.

NOTE: this will just affect the priority of transfers that end up on
the SPI core's pumping thread.  In many cases transfers happen in the
context of the caller so if you need realtime priority for all
transfers you should ensure the calling context is also realtime
priority.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 924b5867e7bd6a6a98014f0517b747465b108011
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/spi.git topic/pump-rt)

Conflicts:
   include/linux/spi/spi.h
...context: "mode" is u16 in our tree and u32 upstream.

BUG=chromium:948742
TEST=Stress writes to /var/log

Change-Id: Iee57becfa790410a1586424463180ae1a1275b39
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1626797
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>