CHROMIUM: iwl7000: mac80211: fill low rate even for HAS_RATE_CONTROL
If HW advertises it has rate control, we skip all of the
rate control assignments, but sometimes the data we have
here is useful, especially so that we don't have to do
the lookups again on which rates are configured and are
supported.

So do the low rate assignment anyway to help out drivers
that might need it.

BUG=b:132825853
TEST=network_WiFi_APSupportedRates

Change-Id: I74d66854d88b73600c0d9628a664a12c06d7965b
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: f141576b10f911ff3cc215cab059effa7c64716c
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1653694
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>