CHROMIUM: mfd: cros_ec: Try loading cros_ec_accel_legacy on more models
Older ARM devices may support acceleromter in legacy mode but the EC
firmare is neither using LPC mode nor listing its features.

All cros-ec-accel-legacy to be loaded to check if there are sensors
available. An error will be reported in the IIO driver
cros-ec-accel-legacy if not.

BUG=chromium:965274
TEST=On veyron minnie, check chrome detect tablet mode and rotate
screen in tablet mode.

Change-Id: I647ea11be5774420dc4a160918dabc017f12fc52
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1632658
Reviewed-by: Douglas Anderson <dianders@chromium.org>