CHROMIUM: virtwl: Kick the in virtqueue in virtwl_vfd_remove()
virtwl_vfd_remove() returns any in-queue buffers owned by the vfd to
the device, but doesn't kick the virtqueue. This works out as long as
the in-queue isn't completely full, because the next thing to be
written to it will trigger a queue kick, but if the queue is
completely blocked the system does not recover. Fix this by adding a
call to virtqueue_kick() to virtwl_vfd_remove()

BUG=chromium:897802
TEST=manually tested

Change-Id: Ic4aa33c69338f0df567827a85b15453bb69584ac
Signed-off-by: Fergus Dall <sidereal@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1627438
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>