CHROMIUM: drm/i915: Disable boost and waitboost
The GT frequency should ideally scale as per the load,
but on making a Hangouts call or opening the camera
preview it is pegged to max frequency. This leads to higher than
expected power consumption(2W+).

The camera stack is probably doing some heavy handed sync operations
which causes the GPU to be always stuck at max frequency.
This is a temporary fix until the userspace(camera) is either
fixed or optimized because we may incur a performance penalty due to
this change.

This bug is discussed here[1].

1. https://bugs.freedesktop.org/show_bug.cgi?id=110750

BUG=b:130638275
TEST=Make hangouts call and check power_supply_info energy rate(W).
     Make sure to disconnect the DUT from AC power.
     With this patch it shows around 2W power savings on average.

Change-Id: I609db0170bdf934d39844078163a55b9368ae3db
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>
Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1629431
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>