FIXUP: FROMLIST: USB: dwc2: Don't turn off the usbphy in suspend if wakeup is enabled
===

This is a revert of v2 of the patch that we landed FROMLIST squashed
with a pick of v3.  See CL:1593684 for v2.

===

If the 'snps,need-phy-for-wake' is set in the device tree then:

- We know that we can wakeup, so call device_set_wakeup_capable().
  The USB core will use this knowledge to enable wakeup by default.
- We know that we should keep the PHY on during suspend if something
  on our root hub needs remote wakeup.  This requires the patch (USB:
  Export usb_wakeup_enabled_descendants()).  Note that we don't keep
  the PHY on at suspend time if it's not needed because it would be a
  power draw.

If we later find some users of dwc2 that can support wakeup without
keeping the PHY on we may want to add a way to call
device_set_wakeup_capable() without keeping the PHY on at suspend
time.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
(am from https://lore.kernel.org/patchwork/patch/1076547/)
(also found at https://lkml.kernel.org/r/20190520175605.2405-3-dianders@chromium.org)

BUG=chromium:941638
TEST=USB wake still works OK

Change-Id: I97d102c1936edeb0de9ea2ae8ec438460e3a1c24
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1632649
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>