FIXUP: CHROMIUM: platform/chrome: wilco_ec: Add charge scheduling sysfs
There were several mistakes in the initial implementation:

The EC property IDs were all messed up in the initial commit.
This corrects them to follow the specification at
https://docs.google.com/document/d/1lNgipPdQJySZ8VIPtM0UO0SGeupBR7Vbdu-obFmZ8mI

The EC interface specification was incorrect in the format for the
Advanced Charging arguments: http://issuetracker.google.com/133749406
This fixes the format.

Also, this has a couple of improvements:
- Refactor ALL of the getters and setters into charge_schedule.h,
  so that charge_schedule_sysfs doesn't need to know about property IDs
- Split up the Advanced Charging and Peak Shift attributes into
  separate directories with the attribute_group struct.
- Naming tweaks and some other cleanup.

BUG=b:132781159, b:133763518, b:133749406
TEST=Using a USB-C power meter, playing with the
policies has an effect on the AC power draw

Change-Id: I253c3516e8beaa050fb842fad0c267735d4174d8
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1633254
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>