UPSTREAM: media: v4l2-mem2mem: add try_en/decoder_cmd ioctl helpers
Most if not all codecs will need to implement these ioctls and
it is expected to be the same for all codecs. So add this to
the core v4l2-mem2mem framework so that this code can easily be
reused.

BUG=b:79450911
BUG=b:123551776
TEST=Build Cheza and Kukui images.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit ef732d5e2813ff5ef30cdb280b2d133e74213555)
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: If5c1a4f4eb88557456884e380e0ddcc52af810b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1786332
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>