CHROMIUM: iwl7000: disable d0i3 by default
d0i3 should be not enabled, certainly not by default.
For an unclear reason, we had a default value for the
module parameter configurable and controlled by a
Kconfig option. This Kconfig option was wrong and because
of that, we didn't disable d0i3.
The d0i3 code should be removed, but as a first immediate
step make sure it is disabled by default.

Change-Id: I9772e784747675fb79d517e25353321e8fa358ca
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: d58910fd72ad0bf51fd003c67277185cfb58edbd